requirements.txt
alabaster==0.7.6
Babel==2.0
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
Cython==0.23
docutils==0.12
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.4.3
mock==1.3.0
nose==1.3.7
osxrelocator==1.0.1
pbr==1.6.0
Pillow==2.9.0
pycrypto==2.6.1
pyenchant==1.6.6
Pygments==2.0.2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
PyX==0.12.1
vboxapi==1.0
